# Optimized attention dot product